fix: put multicodec back to snap
