+++ packages/types/index.ts
@@ -3,5 +3,3 @@ export * from './requests';
 export * from './snapApi';
 export * from './constants';
 export * from './params';
-
-export type Callback<T> = (arg: T) => void;
